Remove unnecessary dashboard components, refactor
