+++ src/pages/common/DashboardPage/DashboardPage.module.less
@@ -1,3 +1,4 @@
+// ** DashboardPage ** //
 .pageWrapper {
   display: flex;
   flex-direction: column;
@@ -6,13 +7,12 @@
 }
 
 .content {
-  flex-grow: 1;
-  overflow-y: scroll;
-  height: 100vh;
-
   display: flex;
-  padding: 0 24px;
+  align-items: center;
+  padding: 24px;
   gap: 24px;
+  max-width: 1240px;
+  margin: 0 auto;
   width: 100%;
 
   @media (max-width: 1240px) {
@@ -24,68 +24,16 @@
     padding: 0 16px;
   }
 }
-
-.searchableHeadingWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 24px 0 12px;
-  gap: 24px;
-  width: 100%;
-
-  @media (max-width: 1440px) {
-    padding: 16px 0 8px;
-    align-items: flex-start;
-    flex-direction: column;
-    gap: 16px;
-  }
-}
-.searchableHeading {
-  font: var(--title-xl);
-  text-transform: uppercase;
-  white-space: nowrap;
-  width: 50%;
-
-  @media (max-width: 640px) {
-    font: var(--title-md);
-    line-height: 20px;
-  }
-}
-
-.searchSelect {
-  width: 50%;
-
-  @media (max-width: 1440px) {
-    width: 100%;
-  }
-}
-
-.searchSelect :global {
-  height: 32px;
-
-  .rootSelectClassName {
-    .ant-select-selector {
-      padding: 1px 0px;
-    }
-  }
-
-  .rootSelectPopupClassName {
-    top: 31px !important;
-  }
-}
+// ** DashboardPage ** //
 
 .heading {
   display: flex;
   align-items: center;
-  padding: 24px 0 12px;
   font: var(--title-xl);
   text-transform: uppercase;
-  line-height: 32px;
 
   @media (max-width: 640px) {
-    padding: 8px 0;
     font: var(--title-md);
-    line-height: 20px;
   }
 }
 
@@ -113,6 +61,7 @@
   }
 }
 
+// * ChartStatInfo * //
 .chartStatInfo {
   display: flex;
   align-items: center;
@@ -147,14 +96,4 @@
   height: 8px;
   width: 8px;
 }
-
-.searchSelectOptionLabel {
-  max-width: 140px;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-
-  @media (max-width: 1440px) {
-    max-width: 100%;
-  }
-}
+// * ChartStatInfo * //
