+++ src/pages/common/DashboardPage/DashboardPage.tsx
@@ -1,45 +1,26 @@
-import { Tab, Tabs, useTabs } from '@banx/components/Tabs'
-
 import { useOnboardingModal } from '@banx/hooks'
 
-import DashboardBorrowTab from './components/DashboardBorrowTab'
+import AllocationBlock from './components/AllocationBlock'
+import MyLoans from './components/MyLoans/components/MyLoans'
 import DashboardHeader from './components/DashboardHeader'
-import DashboardLendTab from './components/DashboardLendTab'
+import { useBorrowerStats, useLenderStats } from './hooks'
 
 import styles from './DashboardPage.module.less'
 
 export const DashboardPage = () => {
   useOnboardingModal('dashboard')
 
-  const { value: currentTabValue, ...tabsProps } = useTabs({
-    tabs: TABS,
-    defaultValue: TabName.BORROW,
-  })
+  const { data: lenderStats } = useLenderStats()
+  const { data: borrowerStats } = useBorrowerStats()
 
   return (
     <div className={styles.pageWrapper}>
       <DashboardHeader />
-      <Tabs value={currentTabValue} {...tabsProps} />
+
       <div className={styles.content}>
-        {currentTabValue === TabName.BORROW && <DashboardBorrowTab />}
-        {currentTabValue === TabName.LEND && <DashboardLendTab />}
+        <MyLoans stats={borrowerStats} />
+        <AllocationBlock stats={lenderStats} />
       </div>
     </div>
   )
 }
-
-enum TabName {
-  BORROW = 'borrow',
-  LEND = 'lend',
-}
-
-const TABS: Tab[] = [
-  {
-    label: 'Borrow',
-    value: TabName.BORROW,
-  },
-  {
-    label: 'Lend',
-    value: TabName.LEND,
-  },
-]
