--- src/pages/common/DashboardPage/components/Card/Card.module.less
@@ -1,81 +1,0 @@
-.card {
-  background-color: var(--bg-primary);
-  border-radius: 6px;
-
-  display: flex;
-  flex-direction: column;
-  padding: 8px;
-  gap: 8px;
-}
-
-.mainCardInfo {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.imageWrapper {
-  position: relative;
-  height: 32px;
-  width: 32px;
-}
-
-.mainCardImage {
-  border-radius: 100px;
-  height: 32px;
-  width: 32px;
-  position: relative;
-  z-index: 2;
-}
-
-.imageShadow {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 32px;
-  width: 32px;
-  border-radius: 100px;
-  filter: blur(6px);
-  z-index: 1;
-  opacity: 0.5;
-  transform: scale(1.1);
-}
-
-.mainCardName {
-  font: var(--title-sm);
-
-  @media (max-width: 640px) {
-    font: var(--title-xs);
-  }
-}
-
-.additionalCardInfo {
-  display: flex;
-  flex-direction: column;
-}
-
-.dashedLine {
-  margin: 4px -8px 8px;
-
-  background: repeating-linear-gradient(
-    to right,
-    var(--bg-border) 0,
-    var(--bg-border) 10px,
-    transparent 10px,
-    transparent 15px
-  );
-
-  height: 1px;
-  width: calc(100% + 16px);
-}
-
-.additionalLabelStat {
-  color: var(--content-tertiary);
-  font: var(--title-xs);
-  line-height: 18px;
-}
-
-.additionalValueStat {
-  font: var(--body-text-sm);
-  line-height: 18px;
-}
