--- src/pages/common/DashboardPage/components/Card/index.ts
@@ -1,1 +1,0 @@
-export * from './Card'
