--- src/pages/common/DashboardPage/components/DashboardBorrowTab/DashboardBorrowTab.module.less
@@ -1,49 +1,0 @@
-.nftsSection,
-.additionalSection {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  width: 50%;
-
-  @media (max-width: 1240px) {
-    flex-grow: unset;
-    width: 100%;
-  }
-}
-
-.additionalSection.fixedHeight > div {
-  flex-grow: unset;
-}
-
-.cardsList {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(180px, 2fr));
-  gap: 24px;
-  margin-top: 12px;
-  padding-bottom: 12px;
-  overflow: auto;
-  z-index: 1; //? prevent bug with SearchSelect overlap on phantom mobile app
-
-  @media (max-width: 1240px) {
-    display: flex;
-
-    & > div {
-      min-width: 180px;
-    }
-  }
-
-  @media (max-width: 640px) {
-    padding-top: 8px;
-    margin-right: -16px;
-    padding-right: 16px;
-    gap: 16px;
-
-    & > div {
-      min-width: 154px;
-    }
-  }
-}
-
-.emptyList {
-  margin: 0;
-}
