--- src/pages/common/DashboardPage/components/DashboardBorrowTab/components/AvailableToBorrow/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AvailableToBorrow'
