--- src/pages/common/DashboardPage/components/DashboardBorrowTab/components/MyLoans/MyLoans.module.less
@@ -1,103 +1,0 @@
-.loansContainer {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-}
-
-.loansContent {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  margin-bottom: 16px;
-  gap: 24px;
-  min-height: 136px;
-  height: 100%;
-  width: 100%;
-}
-
-.loansStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 8px;
-  height: 100%;
-  width: 100%;
-}
-
-.loansStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-
-  @media (max-width: 640px) {
-    margin-top: 8px;
-  }
-}
-
-.totalDebt {
-  flex-direction: column-reverse;
-  white-space: nowrap;
-  gap: 0;
-}
-
-.separateLine {
-  background: var(--bg-border);
-  height: 100%;
-  width: 1px;
-}
-
-.additionalStats {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  white-space: nowrap;
-  width: 100%;
-
-  & > div {
-    justify-content: flex-start;
-    gap: 8px;
-  }
-}
-
-.loansChartStats {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-secondary);
-  }
-}
-
-.desktopChart {
-  @media (max-width: 640px) {
-    display: none;
-  }
-}
-
-.mobileChart {
-  display: none;
-
-  @media (max-width: 640px) {
-    display: flex;
-    justify-content: center;
-  }
-}
-
-.manageLoansButton {
-  margin-bottom: 24px;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin-bottom: 16px;
-  }
-}
-
-.doughnutChart {
-  width: 136px;
-  height: 136px;
-}
-
-.highlightLiquidation {
-  background: var(--additional-red-secondary) !important;
-}
