--- src/pages/common/DashboardPage/components/DashboardBorrowTab/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DashboardBorrowTab'
