--- src/pages/common/DashboardPage/components/DashboardHeader/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './DashboardHeader'
