--- src/pages/common/DashboardPage/components/DashboardLendTab/DashboardLendTab.module.less
@@ -1,45 +1,0 @@
-.section,
-.additionalContentSection {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  width: 50%;
-
-  @media (max-width: 1240px) {
-    flex-grow: unset;
-    width: 100%;
-  }
-}
-
-.section.fullWidth {
-  width: 100%;
-}
-
-.cardsList {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(180px, 2fr));
-  margin-top: 12px;
-  padding-bottom: 12px;
-  gap: 24px;
-  overflow: auto;
-  z-index: 1; //? prevent bug with SearchSelect overlap on phantom mobile app
-
-  @media (max-width: 1240px) {
-    display: flex;
-
-    & > div {
-      min-width: 180px;
-    }
-  }
-
-  @media (max-width: 640px) {
-    padding-top: 8px;
-    margin-right: -16px;
-    padding-right: 16px;
-    gap: 16px;
-
-    & > div {
-      min-width: 154px;
-    }
-  }
-}
