--- src/pages/common/DashboardPage/components/DashboardLendTab/components/AllTimeBlock/AllTimeBlock.module.less
@@ -1,84 +1,0 @@
-.allTimeContainer {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: 100%;
-
-  @media (max-width: 640px) {
-    margin-top: 16px;
-  }
-}
-
-.allTimeContent {
-  display: flex;
-  align-items: flex-end;
-  margin-bottom: 24px;
-  gap: 16px;
-  width: 100%;
-  height: 100%;
-}
-
-.allTimeStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 16px;
-  min-height: 188px;
-  height: 100%;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin-top: 8px;
-  }
-}
-
-.totalLent {
-  flex-direction: column-reverse;
-  white-space: nowrap;
-  gap: 0;
-}
-
-.separateLine {
-  background: var(--bg-border);
-  height: 100%;
-  width: 1px;
-}
-
-.allTimeStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  width: 100%;
-}
-
-.interestStats {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  white-space: nowrap;
-  width: 100%;
-
-  & > div {
-    justify-content: flex-start;
-    gap: 8px;
-  }
-}
-
-.allTimeChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-secondary);
-  }
-}
-
-.singleBarChart {
-  width: 144px;
-  max-height: 188px;
-  height: 100%;
-
-  @media (max-width: 640px) {
-    width: 48px;
-  }
-}
