--- src/pages/common/DashboardPage/components/DashboardLendTab/components/AllTimeBlock/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AllTimeBlock'
