--- src/pages/common/DashboardPage/components/DashboardLendTab/components/AllocationBlock/AllocationBlock.module.less
@@ -1,67 +1,0 @@
-.allocationContainer {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: 100%;
-}
-
-.allocationContent {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  flex-grow: 1;
-  gap: 24px;
-}
-
-.allocationStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 160px;
-  height: 100%;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin-top: 8px;
-  }
-}
-
-.allocationStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.allocationChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-secondary);
-  }
-}
-
-.desktopChart {
-  @media (max-width: 640px) {
-    display: none;
-  }
-}
-
-.mobileChart {
-  display: none;
-
-  @media (max-width: 640px) {
-    display: flex;
-    justify-content: center;
-  }
-}
-
-.doughnutChart {
-  width: 144px;
-  height: 144px;
-}
-
-.manageOffersButton {
-  margin-top: 16px;
-  width: 100%;
-}
