--- src/pages/common/DashboardPage/components/DashboardLendTab/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DashboardLendTab'
